Re-added logging and progress bar

+++ retrieval/pubmed-entrez.py
@@ -1,41 +1,75 @@
-from http.client import IncompleteRead
+from http.client import IncompleteRead, BadStatusLine
 from Bio import Entrez
 import os
 import gzip
+import logging
+import progressbar
+
+
+# set up logging to file - see previous section for more details
+logging.basicConfig(level=logging.DEBUG,
+                    format='%(asctime)s %(name)-12s %(levelname)-8s %(message)s',
+                    datefmt='%m-%d %H:%M',
+                    filename='fetch.log',
+                    filemode='w')
+# define a Handler which writes INFO messages or higher to the sys.stderr
+console = logging.StreamHandler()
+console.setLevel(logging.INFO)
+# set a format which is simpler for console use
+formatter = logging.Formatter('%(name)-12s: %(levelname)-8s %(message)s')
+# tell the handler to use this format
+console.setFormatter(formatter)
+# add the handler to the root logger
+logging.getLogger('').addHandler(console)
+# create console handler with a higher log level
+ch = logging.StreamHandler()
+ch.setLevel(logging.INFO)
 
 BASE_PATH = "/home/alex/medline"
+BATCH_SIZE = 10000
 
 Entrez.email = "apurdy@uvic.ca"
 
-for year in range(2015, 1800, -1):
-    records_dir_path = os.path.join(BASE_PATH, str(year))
-    if not os.path.exists(records_dir_path):
-        os.makedirs(records_dir_path)
+for year in range(2007, 1800, -1):
 
     search_results = Entrez.read(Entrez.esearch(db="pubmed",
                                                 term="{0:d}[edat]".format(year),
                                                 usehistory="y"))
     count = int(search_results["Count"])
-    print("Found %i results" % count)
+    logging.info("Found {0:d} records published in {1:d}".format(count, year))
     if count == 0:
         continue
 
-    batch_size = 10000
-    for start in range(0, count, batch_size):
-        end = min(count, start + batch_size)
-        fname = os.path.join(records_dir_path, "{0:d}-{1:d}.xml.gz".format(start + 1, end))
+    records_dir_path = os.path.join(BASE_PATH, str(year))
+    if not os.path.exists(records_dir_path):
+        os.makedirs(records_dir_path)
 
+    bar = progressbar.ProgressBar(maxval=count - 1,
+                                  widgets=[progressbar.Bar('=', '[', ']'), ' ', progressbar.Percentage()])
+    bar.start()
+    for start in range(0, count, BATCH_SIZE):
+        end = min(count - 1, start + BATCH_SIZE)
+        fname = os.path.join(records_dir_path, "{0:d}-{1:d}.xml.gz".format(start + 1, end))
+        if os.path.exists(fname):
+            bar.update(end)
+            continue
         with gzip.open(fname, "wt") as out_handle:
-            print("Going to download record %i to %i" % (start + 1, end))
-            fetch_handle = Entrez.efetch(db="pubmed",
-                                         retmode="xml",
-                                         retstart=start, retmax=batch_size,
-                                         webenv=search_results["WebEnv"],
-                                         query_key=search_results["QueryKey"])
+            logging.debug("Downloading record {0:d} to {1:d}".format(start + 1, end))
             try:
-                data = fetch_handle.read()
-                out_handle.write(data)
-            except IncompleteRead:
-                print("Fail to read records %i to %i" % (start + 1, end))
-            finally:
-                fetch_handle.close()
+                fetch_handle = Entrez.efetch(db="pubmed",
+                                             retmode="xml",
+                                             retstart=start, retmax=BATCH_SIZE,
+                                             webenv=search_results["WebEnv"],
+                                             query_key=search_results["QueryKey"])
+                try:
+                    data = fetch_handle.read()
+                    out_handle.write(data)
+                except IncompleteRead:
+                    logging.error("Fail to read records {0:d} to {1:d}".format(start + 1, end))
+                finally:
+                    fetch_handle.close()
+            except BadStatusLine:
+                logging.error("Fail to read records {0:d} to {1:d}".format(start + 1, end))
+
+
+        bar.update(end)
